Use static_cast to convert between float/int.
So we can avoid the warnings on possible loss of data when doing
type conversions. They are treated as error on Windows.

+++ cobalt/math/point.h
@@ -23,7 +23,9 @@ class Point : public PointBase<Point, int, Vector2d> {
 
   ~Point() {}
 
-  operator PointF() const { return PointF(x(), y()); }
+  operator PointF() const {
+    return PointF(static_cast<float>(x()), static_cast<float>(y()));
+  }
 
   // Returns a string representation of point.
   std::string ToString() const;
